requirements.txt
asn1crypto==0.22.0
cffi==1.11.0
click==6.7
cryptography==2.0.3
enum34==1.1.6
Flask==0.12.2
Flask-Cache==0.13.1
flask-marshmallow==0.8.0
gitdb2==2.0.0
github3.py==1.0.0a4
GitPython==2.1.0
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
marshmallow==2.13.6
ndg-httpsclient==0.4.2
pkg-resources==0.0.0
pyasn1==0.1.9
pycparser==2.18
PyGithub==1.35
pygithub3==0.5.1
PyJWT==1.5.3
pyOpenSSL==16.2.0
PyYAML==3.12
requests==2.12.1
six==1.11.0
smmap2==2.0.1
uritemplate==3.0.0
uritemplate.py==3.0.2
Werkzeug==0.12.2
